docker/api/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
tensorflow==2.18.0
Pillow==11.1.0
numpy==2.0.2
python-multipart